refactor effects initialization to fix chain routing switches
Initialization of the StandardEffectRack depends on
PlayerManager creating the channels and registering them with
EffectsManager, otherwise EngineEffectChain's
m_chainStatusForChannelMatrix is not initialized properly.
Previously this was hacked around by EffectChain not creating
the EngineEffectChain until it was added to the engine, but
we have removed the possibility of an EffectChainSlot existing
without having an EngineEffectChain added to the engine.