use EffectManifests instead of QString IDs to identify effects
Effect ID strings are only guaranteed to be unique for one
EffectsBackend. Using only the ID strings to identify effects creates
the possibility of a name collision between different effect types from
different EffectsBackends. EffectManifests contain both the ID string
and the EffectBackendType so they uniquely identify an effect type.

This refactoring forced a big cleanup of DlgPrefEq. Lots of duplicate
code was removed from that class, but there is still tight coupling
between DlgPrefEq and EffectChainSlot + EffectSlot for the master
equalizer.